backend/requirements.txt
Flask==2.1.3
requests==2.26.0
pymongo==3.12.0
python-dotenv==0.19.0
Werkzeug==2.1.2
Flask-CORS==3.0.10